im tired

+++ Frontend/src/components/Users.js
@@ -1,36 +1,37 @@
 // UserPage.js
-import React, { useState } from 'react';
+import React, { useState, useEffect, useCallback } from 'react';
 import { Button, Table, TableBody, TableCell, TableContainer, TableHead, TableRow, Paper, Dialog, DialogTitle, DialogContent, DialogActions, TextField } from '@mui/material';
-import Navigation from './Navigation'
-
-// useEffect(() => {
-//   // Fetch users from the backend when the component mounts
-//   fetchUsers();
-// }, []);
-
-// const initialUsers = [
-//   { id: 1, name: 'Ahmad Ali', email: 'ahhmadaliyou2000@gmail.com', Username: 'Hamdi', password: '1234' },
-//   { id: 2, name: 'Hayat Tuge', email: 'hsmltuge@gmail.com', Username: 'Hayat', password: '1234' },
-// ];
+import Navigation from './Navigation';
 
 const UserPage = () => {
-  const [users, setUsers] = useState([]);
+  const [users, setUsers] = useState([
+    { id: 1, name: 'Ahmad Ali', email: 'ahhmadaliyou2000@gmail.com', username: 'Hamdi', password: '12345' }
+  ]);
   const [open, setOpen] = useState(false);
   const [selectedUser, setSelectedUser] = useState({});
   const [name, setName] = useState('');
   const [email, setEmail] = useState('');
-  const [Username, setUsername] = useState('');
+  const [username, setUsername] = useState('');
   const [password, setPassword] = useState('');
 
-  const fetchUsers = async () => {
+  const fetchUsers = useCallback(async () => {
     try {
       const response = await fetch('http://localhost:5000/api/users');
-      const data = await response.json();
-      setUsers(data);
+      if (response.ok) {
+        const data = await response.json();
+        setUsers([users[0], ...data]);
+      } else {
+        console.error('Error fetching users:', response.statusText);
+      }
     } catch (error) {
       console.error('Error fetching users:', error);
     }
-  };
+  }, [users]);
+
+  useEffect(() => {
+    // Fetch other users from the backend when the component mounts
+    fetchUsers();
+  }, [fetchUsers]);
 
   const handleOpen = (user) => {
     setSelectedUser(user);
@@ -61,14 +62,14 @@ const UserPage = () => {
           id: selectedUser.id,
           name,
           email,
-          Username,
+          username,
           password,
         }),
       });
 
       if (response.ok) {
         console.log('User saved successfully');
-        fetchUsers(); // Fetch updated user list
+        fetchUsers();
       } else {
         console.error('Error saving user');
       }
@@ -80,6 +81,12 @@ const UserPage = () => {
   };
 
   const handleDelete = async (id) => {
+    // Prevent deleting the admin user
+    if (id === 1) {
+      console.log('Cannot delete admin user');
+      return;
+    }
+
     try {
       const response = await fetch(`http://localhost:5000/api/users/${id}`, {
         method: 'DELETE',
@@ -87,7 +94,7 @@ const UserPage = () => {
 
       if (response.ok) {
         console.log('User deleted successfully');
-        fetchUsers(); // Fetch updated user list
+        fetchUsers();
       } else {
         console.error('Error deleting user');
       }
@@ -97,88 +104,93 @@ const UserPage = () => {
   };
 
   return (
-   <><Navigation/>
-    <div>
-      <h1>User Page</h1>
-      <Button variant="contained" color="primary" onClick={() => handleOpen({})}>
-        Add User
-      </Button>
-      <TableContainer component={Paper} style={{ marginTop: '16px' }}>
-        <Table>
-          <TableHead>
-            <TableRow>
-              <TableCell>ID</TableCell>
-              <TableCell>Name</TableCell>
-              <TableCell>Email</TableCell>
-              <TableCell>Username</TableCell>
-              <TableCell>Password</TableCell>
-              <TableCell>Action</TableCell>
-            </TableRow>
-          </TableHead>
-          <TableBody>
-            {users.map((user) => (
-              <TableRow key={user.id}>
-                <TableCell>{user.id}</TableCell>
-                <TableCell>{user.name}</TableCell>
-                <TableCell>{user.email}</TableCell>
-                <TableCell>{user.password}</TableCell>
-                <TableCell>
-                  <Button color="primary" onClick={() => handleOpen(user)}>
-                    Edit
-                  </Button>
-                  <Button color="secondary" onClick={() => handleDelete(user.id)}>
-                    Delete
-                  </Button>
-                </TableCell>
+    <>
+      <Navigation />
+      <div>
+        <h1>User Page</h1>
+        <Button variant="contained" color="primary" onClick={() => handleOpen({})}>
+          Add User
+        </Button>
+        <TableContainer component={Paper} style={{ marginTop: '16px' }}>
+          <Table>
+            <TableHead>
+              <TableRow>
+                <TableCell>ID</TableCell>
+                <TableCell>NAME</TableCell>
+                <TableCell>EMAIL</TableCell>
+                <TableCell>USERNAME</TableCell>
+                <TableCell>PASSWORD</TableCell>
+                <TableCell>ACTION</TableCell>
               </TableRow>
-            ))}
-          </TableBody>
-        </Table>
-      </TableContainer>
-      <Dialog open={open} onClose={handleClose}>
-        <DialogTitle>{selectedUser.id ? 'Edit User' : 'Add User'}</DialogTitle>
-        <DialogContent>
-          <TextField
-            label="Name"
-            value={name}
-            onChange={(e) => setName(e.target.value)}
-            fullWidth
-            margin="normal"
-          />
-          <TextField
-            label="Email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
-            fullWidth
-            margin="normal"
-          />
-          <TextField
-            label="Username"
-            value={Username}
-            onChange={(e) => setUsername(e.target.value)}
-            fullWidth
-            margin="normal"
-          />
-          <TextField
-            label="password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
-            fullWidth
-            margin="normal"
-          />
-          
-        </DialogContent>
-        <DialogActions>
-          <Button onClick={handleClose} color="primary">
-            Cancel
-          </Button>
-          <Button onClick={handleSave} color="primary">
-            Save
-          </Button>
-        </DialogActions>
-      </Dialog>
-    </div>
-   </>
+            </TableHead>
+            <TableBody>
+              {users.map((user) => (
+                <TableRow key={user.id}>
+                  <TableCell>{user.id}</TableCell>
+                  <TableCell>{user.name}</TableCell>
+                  <TableCell>{user.email}</TableCell>
+                  <TableCell>{user.username}</TableCell>
+                  <TableCell>{user.password}</TableCell>
+                  <TableCell>
+                    <Button color="primary" onClick={() => handleOpen(user)}>
+                      Edit
+                    </Button>
+                    <Button
+                      color="secondary"
+                      onClick={() => handleDelete(user.id)}
+                      disabled={user.id === 1}
+                    >
+                      Delete
+                    </Button>
+                  </TableCell>
+                </TableRow>
+              ))}
+            </TableBody>
+          </Table>
+        </TableContainer>
+        <Dialog open={open} onClose={handleClose}>
+          <DialogTitle>{selectedUser.id ? 'Edit User' : 'Add User'}</DialogTitle>
+          <DialogContent>
+            <TextField
+              label="Name"
+              value={name}
+              onChange={(e) => setName(e.target.value)}
+              fullWidth
+              margin="normal"
+            />
+            <TextField
+              label="Email"
+              value={email}
+              onChange={(e) => setEmail(e.target.value)}
+              fullWidth
+              margin="normal"
+            />
+            <TextField
+              label="Username"
+              value={username}
+              onChange={(e) => setUsername(e.target.value)}
+              fullWidth
+              margin="normal"
+            />
+            <TextField
+              label="Password"
+              value={password}
+              onChange={(e) => setPassword(e.target.value)}
+              fullWidth
+              margin="normal"
+            />
+          </DialogContent>
+          <DialogActions>
+            <Button onClick={handleClose} color="primary">
+              Cancel
+            </Button>
+            <Button onClick={handleSave} color="primary">
+              Save
+            </Button>
+          </DialogActions>
+        </Dialog>
+      </div>
+    </>
   );
 };
